check consistency of relation
