+++ src/Centreon/Domain/PlatformTopology/Platform.php
@@ -38,12 +38,6 @@ class Platform
     public const TYPE_MBI = 'mbi';
 
     /**
-     * Broker relation types
-     */
-    public const NORMAL_RELATION = 'normal';
-    public const PEER_RETENTION_RELATION = 'peer_retention';
-
-    /**
      * Available server types
      */
     private const AVAILABLE_TYPES = [
@@ -52,14 +46,6 @@ class Platform
         self::TYPE_REMOTE,
         self::TYPE_MAP,
         self::TYPE_MBI
-    ];
-
-    /**
-     * Available relation types
-     */
-    private const AVAILABLE_RELATIONS = [
-        self::NORMAL_RELATION,
-        self::PEER_RETENTION_RELATION
     ];
 
     /**
